frontend finish with full functionality
login, logout, restricted routes all functioning as intended. registration is also functional. logins are persistent using a check against req.session && req.session.user to return a boolean to the FE that indicates if a session is currently active. if so, state will be set as logged in, to display the locked content.